+++ app/services/renewal_reminders/bulk_first_renewal_reminders_service.rb
@@ -2,11 +2,11 @@
 
 module RenewalReminders
 
-  class TemporaryFirstRenewalReminderService < BulkRenewalRemindersEmailService
+  class BulkFirstRenewalRemindersService < BulkRenewalRemindersEmailService
     private
 
     def send_email(registration)
-      TemporaryFirstRenewalReminderEmailService.run(registration: registration)
+      FirstRenewalReminderEmailService.run(registration: registration)
     end
 
     def expires_in_days